+++ frontend/src/pages/HomePage/index.tsx
@@ -8,33 +8,14 @@ import OngoingSchedules from '@/features/shared-schedule/ui/OngoingSchedules';
 import UpcomingSchedules from '@/features/shared-schedule/ui/UpcomingSchedules';
 
 import { containerStyle } from './index.css';
+import UpcomingSection from './UpcomingSection';
 
-const HomePage = () => {
-  const navigate = useNavigate();
-
-  return (
-    <div className={containerStyle}>
-      <Flex
-        direction='column'
-        gap={700}
-        justify='space-between'
-        width='full'
-      >
-        <Flex justify='space-between' width='full'>
-          <Text typo='h2'>다가오는 일정</Text>
-          <Button
-            onClick={() => navigate({ to: '/upcoming-schedule' })}
-            style='borderless'
-          >
-            모두보기
-          </Button>
-        </Flex>
-        <UpcomingSchedules />
-      </Flex>
-      <OngoingSchedules />
-      <FinishedSchedules />
-    </div>
-  );
-};
+const HomePage = () => (
+  <div className={containerStyle}>
+    <UpcomingSection />
+    <OngoingSchedules />
+    <FinishedSchedules />
+  </div>
+);
 
 export default HomePage;
